requirements.txt
langgraph>=0.2.35
langchain>=0.2.16
pydantic>=2.8.2
pyyaml>=6.0.2
rich>=13.7.1